REFACTOR: Split and moved filtering skills files to filter folder
