+++ components/MaterialLists/SkillList.tsx
@@ -2,7 +2,9 @@
 
 import generateUrl from "@/actions/generateUrl";
 import countMaterialsAttributedToSkill from "@/actions/material/countMaterialsAttributedToSkill";
-import groupSkills, { GroupByOptions } from "@/actions/skills/groupSkills";
+import groupSkills, {
+  GroupByOptions,
+} from "@/actions/skills/group/groupSkills";
 import SkillTag from "@/components/Tags/SkillTag";
 import HeadingThree from "@/components/Text/HeadingThree";
 import { Button } from "@/components/shadcn/ui/button";
